Fix warnings (which are now errors) in the Android build.
BUG=None
TEST=


Review URL: http://codereview.chromium.org/9288034

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@119155 0039d316-1c4b-4281-b951-d872f2087c98

+++ net/base/address_list_unittest.cc
@@ -1,5 +1,5 @@
 FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
-// Copyright (c) 2011 The Chromium Authors. All rights reserved.
+// Copyright (c) 2012 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -325,7 +325,7 @@ TEST(AddressListTest, CreateFromIPAddressList) {
   for (size_t i = 0; i < ARRAYSIZE_UNSAFE(tests); ++i) {
     ASSERT_TRUE(next_ai != NULL);
     EXPECT_EQ(tests[i].ai_family, next_ai->ai_family);
-    EXPECT_EQ(tests[i].ai_addrlen, next_ai->ai_addrlen);
+    EXPECT_EQ(tests[i].ai_addrlen, static_cast<size_t>(next_ai->ai_addrlen));
 
     char* ai_addr = reinterpret_cast<char*>(next_ai->ai_addr);
     int rv = memcmp(tests[i].in_addr,
